packages/model-quantization/pyproject.toml
[project]
name = "model-quantization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bhaskar Mahajan", email = "bhaskarmahajan98@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "common",
    "datasets>=4.8.4",
    "huggingface-hub>=0.36.2",
    "llmcompressor>=0.6.0.1",
    "transformers>=4.52.4",
    "python-dotenv",
    "torch",
    "torchvision"
]

[project.scripts]
model-quantization = "model_quantization:main"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
common = { workspace = true }
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
